# docs/swagger.yaml
definitions:
  models.Login_input:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.SignUp_input:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Root route
      tags:
      - root
  /cart:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Clear user's cart
      tags:
      - сart
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get user's cart
      tags:
      - сart
  /cart/{product_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Remove product from cart
      tags:
      - сart
    post:
      consumes:
      - application/json
      parameters:
      - description: ID of the product to add
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Add product to cart
      tags:
      - сart
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: User credentials (username and password)
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.Login_input'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: User login
      tags:
      - auth
  /logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: User logout
      tags:
      - auth
  /products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get list of products
      tags:
      - products
  /sign_up:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data (username, email, password)
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.SignUp_input'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: User registration
      tags:
      - auth
swagger: "2.0"